Move tests to `parameter_hints.rs`
It seems obvious to me that we are testing a feature that "officially" lives here now
